+++ src/protodantic/__init__.py
@@ -3,6 +3,7 @@
 from ._version import __version__
 from .codegen import generate_source, generate_tree
 from .compiler import compile_fdset
+from .reflection import fdset_from_package
 from .runtime import OpenEnum, ProtoModel, load_pool, model_for
 from .types import NULL, Int32, Int64, ListValue, Struct, UInt32, UInt64, Value
 
@@ -19,6 +20,7 @@ __all__ = [
     "Value",
     "__version__",
     "compile_fdset",
+    "fdset_from_package",
     "generate_source",
     "generate_tree",
     "load_pool",